Pre opt ray's error handlings
