+++ src/main.rs
@@ -37,5 +37,7 @@ pub async fn main() -> Result<()> {
     // let result = extractor_runner(&warc_name).await?;
     // println!("fff{:?}", result);
 
+
+
     Ok(())
 }
